+++ src/App.jsx
@@ -2,17 +2,18 @@ import React from 'react';
 import { Route, Switch, Redirect } from 'react-router';
 import { BrowserRouter } from 'react-router-dom';
 import { hot } from 'react-hot-loader';
-import * as Pages from './pages';
+import * as Pages from './pages/Pages';
 
 const App = () => (
   <BrowserRouter>
     <>
       <Switch>
-        <Route exact path="/" component={Pages.Landing} />
+        <Route exact path="/" component={() => <Redirect to="/login" />} />
         <Route path="/login" component={() => <Pages.Auth type="login" />} />
         <Route path="/signup" component={() => <Pages.Auth type="signup" />} />
         <Route path="/dashboard" component={Pages.Dashboard} />
-        <Route path="*" component={() => <Redirect to="/dashboard" />} />
+        {/* TODO: Make a 404 Page */}
+        <Route path="*" component={() => <Redirect to="/login" />} />
       </Switch>
     </>
   </BrowserRouter>
