+++ src/components/Tear/Tear.scss
