+++ src/pages/Dashboard/Course/Professor/Assignments/Assignments.jsx
@@ -3,13 +3,39 @@ import { Table, Divider, Button, Icon, Modal } from 'antd';
 import { withRouter } from 'react-router';
 import dayjs from 'dayjs';
 
-import Form from 'Components/Form';
+import Form from 'Components/Form/Form';
 import tyr from 'Utils/tyr';
-import { splitAssignments } from './utils';
-import { assignmentForm } from './data';
-import './styles.scss';
+import { assignmentForm } from './formTemplate';
+import './Assignments.scss';
 
 /* eslint-disable no-console */
+
+const sortAssignments = (a, b, asc = true) => {
+  const timeA = dayjs(a.dueDate);
+  const timeB = dayjs(b.dueDate);
+  if (timeA.isSame(timeB)) return 0;
+  return timeA[asc ? 'isBefore' : 'isAfter'](timeB) ? 1 : -1;
+};
+
+const splitAssignments = assignments => {
+  if (assignments.length === 0) {
+    return null;
+  }
+
+  const pastAssignments = [];
+  const upcomingAssignments = [];
+  assignments.sort((a, b) => sortAssignments(a, b, false));
+  assignments.forEach(a => {
+    const dueDate = dayjs(a.dueDate);
+    // this assignment is a past assignment
+    if (dueDate.isBefore(dayjs())) {
+      pastAssignments.push(a);
+    } else {
+      upcomingAssignments.push(a);
+    }
+  });
+  return { upcomingAssignments, pastAssignments };
+};
 
 class Assignments extends Component {
   // reference: https://ant.design/components/table/
