+++ src/pages/Dashboard/Course/Professor/Professor.jsx
@@ -2,19 +2,19 @@ import React, { Component } from 'react';
 
 import { Switch, Route } from 'react-router';
 import { withRouter, Redirect } from 'react-router-dom';
-import Card from 'Components/Card';
+import Card from 'Components/Card/Card';
 import { stringToColour } from 'Utils/misc';
-import Menu from './Menu';
-import Assignments from './Assignments';
-import People from './People';
+import Menu from './Menu/Menu';
+import Assignments from './Assignments/Assignments';
+import People from './People/People';
 
-import './styles.scss';
+import './Professor.scss';
 
 class Professor extends Component {
   state = {};
 
   render() {
-    const { data: course, match } = this.props;
+    const { course, match } = this.props;
     return (
       <div className="professor-view">
         <Card>
