+++ src/pages/Dashboard/Course/Professor/Professor.scss
