+++ src/pages/Dashboard/Course/Student/Student.jsx
@@ -1,0 +1,15 @@
+/* eslint-disable no-underscore-dangle */
+import React, { Component } from 'react';
+
+import './Student.scss';
+
+class Student extends Component {
+  state = {};
+
+  render() {
+    const { course } = this.props;
+    return <pre>{JSON.stringify(course)}</pre>;
+  }
+}
+
+export default Student;
