+++ src/pages/Dashboard/Course/Student/Student.scss
@@ -1,0 +1,3 @@
+@import 'Style/colors.scss';
+@import 'Style/typography.scss';
+@import 'Style/space.scss';
