+++ src/pages/Pages.jsx
@@ -1,0 +1,2 @@
+export { default as Dashboard } from './Dashboard/Dashboard';
+export { default as Auth } from './Auth/Auth';
