--- src/pages/dash/SingleCourse/Collapse/styles.scss
@@ -1,70 +1,0 @@
-@import 'Style/colors.scss';
-@import 'Style/typography.scss';
-@import 'Style/input.scss';
-@import 'Style/shadows.scss';
-
-.collapse {
-  margin: 1rem;
-  background-color: $bg;
-
-  // expanded collapse item
-  .ant-collapse-item > .ant-collapse-content {
-    margin-top: 30px;
-    padding: 15px;
-
-    .panel-body {
-      border-left: 3px solid $grey;
-      padding-left: 20px;
-
-      .required-lang {
-        margin-top: 10px;
-        font-weight: $medium;
-      }
-
-      // panel description header
-      .panel-label {
-        font-size: 1.1em;
-        font-weight: $medium;
-        margin: 10px 0;
-      }
-
-      // center submit button
-      .button-container {
-        margin-top: 1rem;
-        display: flex;
-        justify-content: center;
-      }
-    }
-  }
-
-  .panel {
-    // custom collapse css
-    padding: 2rem;
-    border-radius: 10px;
-    background-color: $white;
-    box-shadow: $raised;
-    margin-bottom: 20px;
-
-    .panel-header {
-      display: flex;
-      justify-content: space-between;
-      margin-right: 30px;
-    }
-
-    .ant-collapse-header {
-      .assignment-name {
-        font-size: 1.2rem;
-      }
-
-      // Line up header text with description text
-      padding-left: 16px;
-      font-weight: $medium;
-
-      .anticon {
-        // Put arrows on right
-        left: initial;
-        right: 16px;
-      }
-    }
-  }
-}
