--- src/pages/index.jsx
@@ -1,4 +1,0 @@
-/* eslint-disable import/no-cycle */
-export { default as Dashboard } from './dash';
-export { default as Landing } from './landing';
-export { default as Auth } from './auth';
