--- src/pages/landing/styles.scss
@@ -1,25 +1,0 @@
-@import 'Style/colors.scss';
-@import 'Style/typography.scss';
-@import 'Style/input.scss';
-
-.landing-container {
-  background-color: $purp;
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-
-  a {
-    width: 10rem;
-    margin-top: 1rem;
-  }
-
-  .logo {
-    vertical-align: middle;
-    display: inline-block;
-    height: 11rem;
-    fill: $white;
-  }
-}
